requirements.txt
Flask==2.3.3
twilio==9.1.0
gunicorn==22.0.0
python-dotenv==1.0.0
